# models/cdd_sdk/generated/cdd_sdkgo/api/openapi.yaml
openapi: 3.0.2
info:
  title: CddService
  version: "1.0"
servers:
- url: /
paths:
...
  /deprovision:
    put:
      operationId: Deprovision
      parameters:
      - explode: true
        in: query
        name: host_id
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: force
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeprovisionResponseContent"
          description: Deprovision 200 response
...
components:
  schemas:
...
    DeprovisionResponseContent:
      example:
        success: true
        state: state
        message: message
        error:
          details: details
          type: type
          message: message
      properties:
        success:
          type: boolean
        state:
          type: string
        message:
          type: string
        error:
          $ref: "#/components/schemas/ErrorDetails"
      required:
      - message
      - state
      - success
      type: object
...
    ErrorDetails:
      example:
        details: details
        type: type
        message: message
      properties:
        type:
          type: string
        message:
          type: string
        details:
          type: string
      required:
      - details
      - message
      - type
      type: object
...